+++ src/main/java/roomescape/application/ReservationService.java
@@ -6,14 +6,14 @@ import java.time.LocalDateTime;
 import java.time.LocalTime;
 import java.util.List;
 import org.springframework.stereotype.Service;
-import roomescape.application.dto.ReservationCreationRequest;
+import roomescape.application.dto.request.ReservationCreationRequest;
+import roomescape.application.dto.response.ReservationResponse;
 import roomescape.domain.reservation.Reservation;
 import roomescape.domain.reservation.repository.ReservationRepository;
 import roomescape.domain.theme.Theme;
 import roomescape.domain.theme.repository.ThemeRepository;
 import roomescape.domain.time.ReservationTime;
 import roomescape.domain.time.repository.ReservationTimeRepository;
-import roomescape.dto.reservation.ReservationResponse;
 
 @Service
 public class ReservationService {
