+++ src/main/java/roomescape/application/dto/request/ReservationTimeCreationRequest.java
@@ -1,4 +1,4 @@
-package roomescape.application.dto;
+package roomescape.application.dto.request;
 
 import java.time.LocalTime;
 import roomescape.domain.time.ReservationTime;
