+++ src/main/java/roomescape/application/dto/request/ThemeCreationRequest.java
@@ -1,4 +1,4 @@
-package roomescape.application.dto;
+package roomescape.application.dto.request;
 
 import roomescape.domain.theme.Theme;
