+++ src/main/java/roomescape/application/dto/response/AvailableTimeResponse.java
@@ -1,4 +1,4 @@
-package roomescape.dto.reservationtime;
+package roomescape.application.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalTime;
