+++ src/main/java/roomescape/controller/ReservationController.java
@@ -16,8 +16,8 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.application.ReservationService;
-import roomescape.dto.reservation.ReservationRequest;
-import roomescape.dto.reservation.ReservationResponse;
+import roomescape.application.dto.response.ReservationResponse;
+import roomescape.controller.dto.ReservationRequest;
 
 @RestController
 @RequestMapping("/reservations")
