+++ src/main/java/roomescape/controller/dto/ReservationRequest.java
@@ -1,10 +1,10 @@
-package roomescape.dto.reservation;
+package roomescape.controller.dto;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 import java.time.LocalDate;
-import roomescape.application.dto.ReservationCreationRequest;
+import roomescape.application.dto.request.ReservationCreationRequest;
 
 public record ReservationRequest(
         @NotBlank(message = "예약자명은 필수입니다.")
