+++ src/main/java/roomescape/controller/dto/ReservationTimeRequest.java
@@ -1,8 +1,8 @@
-package roomescape.dto.reservationtime;
+package roomescape.controller.dto;
 
 import jakarta.validation.constraints.NotNull;
 import java.time.LocalTime;
-import roomescape.application.dto.ReservationTimeCreationRequest;
+import roomescape.application.dto.request.ReservationTimeCreationRequest;
 
 public record ReservationTimeRequest(@NotNull(message = "시작 시간은 필수입니다.") LocalTime startAt) {
     public ReservationTimeCreationRequest toReservationCreationRequest() {
