+++ src/test/java/roomescape/application/ReservationServiceTest.java
@@ -10,12 +10,12 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.dao.DuplicateKeyException;
-import roomescape.application.dto.ReservationCreationRequest;
+import roomescape.application.dto.request.ReservationCreationRequest;
+import roomescape.application.dto.response.ReservationResponse;
 import roomescape.domain.theme.Theme;
 import roomescape.domain.theme.repository.ThemeRepository;
 import roomescape.domain.time.ReservationTime;
 import roomescape.domain.time.repository.ReservationTimeRepository;
-import roomescape.dto.reservation.ReservationResponse;
 import roomescape.fixture.ReservationFixture;
 import roomescape.fixture.ThemeFixture;
 import roomescape.support.annotation.FixedClock;
