+++ src/test/java/roomescape/application/ThemeServiceTest.java
@@ -11,8 +11,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.dao.DuplicateKeyException;
 import org.springframework.test.context.jdbc.Sql;
-import roomescape.application.dto.ThemeCreationRequest;
-import roomescape.dto.theme.ThemeResponse;
+import roomescape.application.dto.request.ThemeCreationRequest;
+import roomescape.application.dto.response.ThemeResponse;
 import roomescape.support.annotation.FixedClock;
 import roomescape.support.extension.MockClockExtension;
 import roomescape.support.extension.TableTruncateExtension;
