+++ src/test/java/roomescape/controller/ReservationTimeControllerTest.java
@@ -16,8 +16,8 @@ import org.junit.jupiter.params.provider.ValueSource;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.web.servlet.ResultActions;
 import roomescape.application.ReservationTimeService;
-import roomescape.dto.reservationtime.ReservationTimeRequest;
-import roomescape.dto.reservationtime.ReservationTimeResponse;
+import roomescape.application.dto.response.ReservationTimeResponse;
+import roomescape.controller.dto.ReservationTimeRequest;
 import roomescape.fixture.ReservationFixture;
 import roomescape.support.ControllerTest;
 import roomescape.support.SimpleMockMvc;
