+++ src/test/java/roomescape/controller/ThemeControllerTest.java
@@ -18,9 +18,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.web.servlet.ResultActions;
 import roomescape.application.ReservationTimeService;
 import roomescape.application.ThemeService;
-import roomescape.dto.reservationtime.AvailableTimeResponse;
-import roomescape.dto.theme.ThemeRequest;
-import roomescape.dto.theme.ThemeResponse;
+import roomescape.application.dto.response.AvailableTimeResponse;
+import roomescape.application.dto.response.ThemeResponse;
+import roomescape.controller.dto.ThemeRequest;
 import roomescape.fixture.ThemeFixture;
 import roomescape.support.ControllerTest;
 import roomescape.support.SimpleMockMvc;
